+++ elements/src/apps/pos-app-rdf-document/pos-app-rdf-document.tsx
@@ -16,6 +16,7 @@ export class PosAppRdfDocument {
               <ion-card-header>
                 <ion-card-title>
                   <pos-label />
+                  <pos-type-badges />
                 </ion-card-title>
                 <pos-literals />
               </ion-card-header>