+++ elements/src/components/pos-type-badges/pos-type-badges.spec.tsx
@@ -1,0 +1,84 @@
+jest.mock('@pod-os/core', () => ({}));
+
+import { newSpecPage } from '@stencil/core/testing';
+
+import { PosTypeBadges } from './pos-type-badges';
+
+describe('pos-type-badges', () => {
+  it('are empty initially', async () => {
+    const page = await newSpecPage({
+      components: [PosTypeBadges],
+      html: `<pos-type-badges />`,
+    });
+    expect(page.root).toEqualHtml(`
+      <pos-type-badges>
+        <mock:shadow-root>
+          <div class="types">
+          </div>
+        </mock:shadow-root>
+      </pos-type-badges>
+  `);
+  });
+
+  it('renders single type', async () => {
+    const page = await newSpecPage({
+      components: [PosTypeBadges],
+      html: `<pos-type-badges />`,
+    });
+    await page.rootInstance.receiveResource({
+      types: () => [
+        {
+          uri: 'https://vocab.test/SomeType',
+          label: 'SomeType',
+        },
+      ],
+    });
+    await page.waitForChanges();
+
+    expect(page.root).toEqualHtml(`
+      <pos-type-badges>
+        <mock:shadow-root>
+          <div class="types">
+            <ion-badge>SomeType</ion-badge>
+          </div>
+        </mock:shadow-root>
+      </pos-type-badges>
+  `);
+  });
+
+  it('renders multiple types', async () => {
+    const page = await newSpecPage({
+      components: [PosTypeBadges],
+      html: `<pos-type-badges />`,
+    });
+    await page.rootInstance.receiveResource({
+      types: () => [
+        {
+          uri: 'https://vocab.test/SomeType',
+          label: 'SomeType',
+        },
+        {
+          uri: 'https://vocab.test/SecondType',
+          label: 'SecondType',
+        },
+        {
+          uri: 'https://vocab.test/AnotherType',
+          label: 'AnotherType',
+        },
+      ],
+    });
+    await page.waitForChanges();
+
+    expect(page.root).toEqualHtml(`
+      <pos-type-badges>
+        <mock:shadow-root>
+          <div class="types">
+            <ion-badge>SomeType</ion-badge>
+            <ion-badge>SecondType</ion-badge>
+            <ion-badge>AnotherType</ion-badge>
+          </div>
+        </mock:shadow-root>
+      </pos-type-badges>
+  `);
+  });
+});